Expose the `insta_science.ensure_installed` API.
Re-structure code to clearly delineate the public APIs through
`insta_science` package re-exports of `insta_science._internal*`
symbols.
