+++ python/CHANGES.md
@@ -1,5 +1,10 @@
 # insta-science
 
+## 0.2.0
+
+Expose the `insta_science.ensure_installed` API for programmatic access
+to the `science` binary's path.
+
 ## 0.1.0
 
 Initial release.
