+++ python/insta_science/_internal/__init__.py
@@ -1,16 +1,11 @@
 # Copyright 2024 Science project contributors.
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from ._internal import (
-    Digest,
-    Fingerprint,
-    InputError,
-    Platform,
-    Science,
-    ScienceNotFound,
-    ensure_installed,
-)
-from .version import __version__
+from .errors import InputError, ScienceNotFound
+from .hashing import Fingerprint
+from .model import Digest, Science
+from .platform import Platform
+from .science import ensure_installed
 
 __all__ = (
     "Digest",
@@ -19,6 +14,5 @@ __all__ = (
     "Platform",
     "Science",
     "ScienceNotFound",
-    "__version__",
     "ensure_installed",
 )
