+++ python/insta_science/version.py
@@ -1,7 +1,4 @@
 # Copyright 2024 Science project contributors.
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from . import shim
-
-if __name__ == "__main__":
-    shim.main()
+__version__ = "0.2.0"
